using migrate from db

+++ app/model/detail_user.py
@@ -3,7 +3,7 @@ from app.model.users import Users
 
 class Detail_User(db.Model):
     id_detail_user = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
-    age = db.Column(db.Integer(50), nullable=False)
+    age = db.Column(db.BigInteger, nullable=False)
     genre = db.Column(db.Enum('Male', 'Female'), nullable=False)
     internet = db.Column(db.Enum('Yes', 'No'), nullable=False)
     fjob = db.Column(db.String(50), nullable=False)
