+++ app/model/evaluation.py
@@ -5,7 +5,7 @@ from app.model.course_subject import Course_Subject
 class Evaluation(db.Model):
     id_evaluation = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
     id_user = db.Column(db.BigInteger, db.ForeignKey(Users.id_user))
-    grade = db.Column(db.Integer(50), nullable=False)
+    grade = db.Column(db.Integer, nullable=False)
     id_course = db.Column(db.BigInteger, db.ForeignKey(Course_Subject.id_course))
